QEDevents

Documentation for QEDevents.

QEDevents.MaxwellBoltzmann

QEDevents.MaxwellBoltzmannParticle

QEDevents._weight

QEDevents.MaxwellBoltzmann — Type

MaxwellBoltzmann(scale::Real)

The Maxwell-Boltzmann distribution with scale parameter a has the probability density function

\[f(x,a) = \sqrt{\frac{2}{\pi}}\frac{x^2}{a^3}\exp\left(\frac{-x^2}{2a^2}\right)\]

The Maxwell-Boltzmann distribution is related to the Chi distribution via the property $X\sim \operatorname{MaxwellBoltzmann}(a=1)$, then $X\sim\chi(\mathrm{dof}=3)$.

External links

Maxwell-Boltzmann distribution on Wikipedia

Maxwell-Boltzmann distribution on Wolfram MathWorld

Maxwell-Boltzmann distribution implementation in Scipy

QEDevents.MaxwellBoltzmannParticle — Type

MaxwellBoltzmannParticle( dir::ParticleDirection, part::AbstractParticleType, temperature::Real )

The Maxwell-Boltzmann particle distribution is a single-particle distribution, where the spatial components of the four-momentum are normally distributed with localion $\mu = 0$ and variance $\sigma = m k_B T$ ($m$ is the particle's mass, $k_B$ is the Boltzmann constant, and $T$ is the temperature). The three-magnitude $\varrho = \sqrt{p_x^2 + p_y^2 + p_z^2}$ of the generated particle-statefuls is MaxwellBoltzmann distributed.

External links

Maxwell-Boltzmann distributed four-momenta on Wikipedia

QEDevents._weight — Method

_weight(
    d::MaxwellBoltzmannParticle{D,P,T}, ps::ParticleStateful{D,P}
) where {D,P,T}

Unsafe weight-function for MaxwellBoltzmannParticle, which is given by

\[ w(p) = \begin{cases} \frac{1}{4\pi}\sqrt{\frac{2}{\pi}}\frac{\varrho^2}{a^3}\exp\left(\frac{-\varrho^2}{2a^2}\right)\quad &\text{,for } p^2=m^2\\ 0 &\mathrm{elsewhere}. \end{cases} \]

with $\varrho^2 = p_x^2 + p_y^2 + p_z^2$ and $a = \sqrt{m k_B T}$ ($m$ is the particle's mass, $k_B$ is the Boltzmann constant, and $T$ is the temperature).