QEDevents

Documentation for QEDevents.

QEDevents.AbstractSampler

Distributions._rand!

QEDevents._assert_valid_input

QEDevents._weight

QEDevents.is_exact

QEDevents.max_weight

QEDevents.setup

QEDevents.weight

QEDevents.AbstractSampler — Type

Abstract base type for general sampler.

Sampler interface

Functions to be implemented

CustomSampler <: AbstractSampler{Float64}

Base.eltype  # type of the generated samples
QEDevents._weight
Base.size  # Return the dimensionality of the sampler (number of outgoing four-momenta per sample).
is_exact
QEDevents._rand!(rng::AbstractRNG, ::AbstractSampler, x::AbstractVector{T}) where {T}

optional

_post_processing

Distributions._rand! — Method

_rand!(
    rng::Random.AbstractRNG,
    smplr::AbstractSampler,
    x::AbstractVector
)

Generate a random sample from the sampler and write it into the given vector.

QEDevents._assert_valid_input — Method

Interface function, which asserts that the given input is valid.

QEDevents._weight — Function

_weight(::AbstractSampler, sample)

Interface function, which returns the weight associated with the given sample according to the sampler.

This function must not do input validation. This is done by weight, which calls _weight after input validation.

QEDevents.is_exact — Function

is_exact(::AbstractSampler)

Interface function, which returns whether the sampler is exactly representing the base distribution or not.

QEDevents.max_weight — Function

max_weight(::AbstractSampler)

Interface function, which returns the maximum possible weight for the sampler.

QEDevents.setup — Function

setup(::AbstractSampler)

Interface function, which returns the setup::AbstractComputeSetup associated with the sampler.

QEDevents.weight — Method

weight(smplr::AbstractSampler, sample)

Interface function, which validates the input, calculates the weight via _weight, and performs an optional post-processing via QEDprocesses._post_processing.