QEDevents

Documentation for QEDevents.

QEDevents.MultiParticleVariate

QEDevents.ParticleLikeVariate

QEDevents.ParticleSampleable

QEDevents.ProcessLikeVariate

QEDevents.SingleParticleDistribution

QEDevents.SingleParticleVariate

QEDevents._assert_valid_input

QEDevents._assert_valid_input_type

QEDevents._assert_valid_input_type

QEDevents._momentum_type

QEDevents._particle

QEDevents._particle_direction

QEDevents._post_processing

QEDevents._weight

QEDevents.is_exact

QEDevents.max_weight

QEDevents.weight

QEDevents.MultiParticleVariate — Type

MultiParticleVariate

Auxiliary type to represent multiple particles in the context of distributions and sampler. A sample from this variate form has the type Vector{ParticleStateful}.

QEDevents.ParticleLikeVariate — Type

ParticleLikeVariate{N<:Int}

Auxiliary type to represent the number of axes in the space of all particles:

N == 0 single particle variate

N == 1 multiple particle variate

QEDevents.ParticleSampleable — Type

ParticleSampleable{V<:QEDlikeVariate}

Abstract base type for sampleable particle distributions and sampler in the context of QEDevents.jl. Here a particle-sampleable is generally a type which has some sort of rand function to produce random samples for properties of given particles, like four-momentum, spin, polarization, etc..

To implement the particle-sampleable interface, the following functions need to be given:

Base.eltype(s::ParticleSampleable): return the innermost type of the samples

_weight(s::ParticleSampleable,x): return the weight of a given sample x

is_exact(s::ParticleSampleable): return wether or not a particle-sampleable is exact

Optionally, one can enhance the calculation of weights by providing

_assert_valid_input_type(s::ParticleSampleable,x): assert input has the correct type

_assert_valid_input(s::ParticleSampleable,x): assert input has correct properties

_post_processing(s::ParticleSampleable,x,res): apply some postprocesing to the result of _weight.

See weight for details. Furthermore, one can provide custom four-momentum types used for the generation of samples by implementing

_momentum_type(s::ParticleSampleable): return the momentum type used

The actual sampling function depends on the variate form:

Single particle sampler

For SingleParticleVariatesamplers, one needs to implement the single sample version rand:

Distributions.rand(
    rng::Random.AbstractRNG,
    s::ParticleSampleable{SingleParticleVariate})

which returns a random sample from s as a ParticleStateful.

Multiple particle sampler

For MultiParticleVariate samplers, one needs to implement

Base.size(s::ParticleSampleable{MultiParticleVariate})

and the mutating version of rand:

Distributions._rand!(
    rng::Random.AbstractRNG,
    s::ParticleSampleable{MultiParticleVariate},
    out::AbstractArray{ParticleStateful})

Scattering process sampler

For ProcessLikeVariate samplers, one needs to give the single sampler version of rand:

Distributions.rand(
    rng::Random.AbstractRNG,
    s::ParticleSampleable{SingleParticleVariate})

which must a random sample from s as a PhaseSpacePoint of the respective scattering process.

QEDevents.ProcessLikeVariate — Type

ProcessLikeVariate

Auxiliary type to represent distributions and samplers for scattering processes.

A sample from this variate form has the type PhaseSpacePoint.

QEDevents.SingleParticleDistribution — Type

SingleParticleDistribution

Base type for sample drawing from single particle distributions. The following interface functions should be implemented:

QEDevents._particle(d::SingleParticleDistribution): return associated particle

QEDevents._particle_direction(d::SingleParticleDistribution): return associated particle direction

Distributions.rand(rng::AbstractRNG,d::SingleParticleDistribution): return random sample as a ParticleStateful

QEDevents.SingleParticleVariate — Type

SingleParticleVariate

Auxiliary type to represent single particles in the context of distributions and sampler. A sample from this variate form has the type ParticleStateful.

QEDevents._assert_valid_input — Method

_assert_valid_input(s::ParticleSampleable,x)

Throw InvalidInputError if the input x is not valid, do nothing otherwise. This function is usually used for checking, if the input has the assumed properties. Type checks are done using multiple dispatch, or _assert_valid_input_type. The default is doing nothing.

Note

This interface function is optional for subtypes of ParticleSampleable.

QEDevents._assert_valid_input_type — Method

_assert_valid_input_type(s::ParticleSampleable,x)

Throw InvalidInputError if the input x has the wrong type, do nothing otherwise. This function is usually used for complicated types, where the implementation via multiple dispatch is cumbersome. The default is doing nothing.

Note

This interface function is optional for subtypes of ParticleSampleable.

QEDevents._assert_valid_input_type — Method

Interface function, which asserts that the given input is valid.

QEDevents._momentum_type — Method

_momentum_type(s::ParticleSampleable,x)

Return the momentum type used for the generation of samples. The default is SFourMomentum.

Note

This interface function is optional for subtypes of ParticleSampleable.

QEDevents._particle — Function

_particle(dist::SingleParticleDistribution)::QEDbase.AbstractParticle

Return the particle associated with the dist.

Note

Interface function to be implemented for single-particle distributions.

QEDevents._particle_direction — Function

_particle_direction(dist::SingleParticleDistribution)::QEDbase.ParticleDirection

Return the particle-direction of the particle associated with dist.

Note

Interface function to be implemented for single-particle distributions.

QEDevents._post_processing — Method

_post_processing(s::ParticleSampleable, x, result)

Return post-processed version of result. The default does nothing and returns result.

Note

This interface function is optional for subtypes of ParticleSampleable.

QEDevents._weight — Function

_weight(s::ParticleSampleable, x)

Return the weight associated with the given sample x according to the particle-sampleable s. This function must not do input validation. This is done by weight, which calls _weight after input validation.

Note

This interface function must be implemented for subtypes of ParticleSampleable.

QEDevents.is_exact — Function

is_exact(s::ParticleSampleable)

Return whether or not the particle-sampleable s is exactly representing the distribution given by weight.

Note

This interface function must be implemented for subtypes of ParticleSampleable.

QEDevents.max_weight — Function

max_weight(::ParticleSampleable)

Interface function, which returns the maximum possible weight for the particle-sampleable.

QEDevents.weight — Method

weight(d::ParticleSampleable, sample)

Return the weight of the given sample according to the given distribution.

This function automatically performs input validation and post-processing using the respective interface functions. The order of calls is

_assert_valid_input_type

_assert_valid_input

_weight

_post_processing